bcvencrypt changes
